BACKPORT: ASoC: core: Allow snd_soc_update_bits use 32 bits register
Change reg's type from unsigned short to unsigned int. So that we can use
32 bits reg value in snd_soc_update_bits.

(cherry picked from commit 1a39019e939f620f39a1b914231ab6ba9013b208)

BUG=chrome-os-partner:29649
TEST=With entire patch series applied, audio playback is working

Conflicts:
	sound/soc/soc-core.c

Reviewed-on: https://chromium-review.googlesource.com/215215
Commit-Queue: Ben Zhang <benzh@chromium.org>
Tested-by: Ben Zhang <benzh@chromium.org>

+++ sound/soc/soc-core.c
@@ -2696,7 +2696,7 @@ int snd_soc_put_volsw_sx(struct snd_kcontrol *kcontrol,
 	int min = mc->min;
 	int mask = (1 << (fls(min + max) - 1)) - 1;
 	int err = 0;
-	unsigned short val, val_mask, val2 = 0;
+	unsigned int val, val_mask, val2 = 0;
 
 	val_mask = mask << shift;
 	val = (ucontrol->value.integer.value[0] + min) & mask;
